// src/types.ts
export interface IUser {
  id: number;
  name: string;
  username: string;
  email: string;
  address: {
    street: string;
    suite: string;
    city: string;
    zipcode: string;
    geo: {
      lat: string;
      lng: string;
    };
  };
  phone: string;
  website: string;
  company: {
    name: string;
    catchPhrase: string;
    bs: string;
  };
}

export interface IApiErrorDetail {
  field: string;
  message: string;
}

export interface IApiError {
  code: number;
  status: string;
  data: IApiErrorDetail[];
}